Basic Javascript implementation of a Trie
To find all words on a board, it will be efficient to check if a given
partial word is a prefix of an existing word in the dictionary. For that
end we use a Trie data structure.
A Trie object is created with the usual Javascript method, new Trie(),
with no arguments. Words are added to the Trie with the insert method.
Lookup is performed with the lookup method. This method takes a string
and returns true if it matches a word in the dictionary, false if it
does not match a word but is a prefix of an existing word and undefined
if it doesn't match and is not a prefix. This behaviour might change.