Added code for solving the puzzle
The code in ruzzle.js finds all words on the board, and the code in
index.js puts them in the wordlist. jQuery is used to get the dictionary
and show and hide messages. The first half of the solve method traverses
the board to find the solutions, and the second half makes them unique
by selecting those with the highest score (as of yet unimplemented).
This needs to be seriously cleaned up.